+++ packages/figma-componentgen-plugin/src/pages/main/MainPage.tsx
@@ -1,0 +1,5 @@
+import { MainContainer } from "./MainContainer";
+
+export const MainPage = () => {
+  return <MainContainer />;
+};
