+++ packages/figma-componentgen-plugin/src/pages/main/component/publish/exception/PublishException.ts
