+++ packages/figma-componentgen-plugin/src/pages/main/component/publish/service/PublishService.ts
