+++ packages/figma-componentgen-plugin/src/pages/main/component/setting/exception/SettingException.ts
