+++ packages/figma-componentgen-plugin/src/pages/main/component/setting/service/MappingTableService.spec.ts
